added helmet. updated readme and changelog.
