+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## v0.09 (dev)
 * Added authentication middleware to router.
 * Added gzip compression.
-* Added PM2 to dockerfile.
+* Added PM2.
+* Added Helmet.
+* Fixed missing session data.
+* Reduced sqlite queries.
 
 ## v0.08 (Dec 15th 2023)
 * Updates to compose file and instructions from [steveiliop56](https://github.com/steveiliop56)
