+++ package.json
@@ -14,6 +14,7 @@
     "ejs": "^3.1.9",
     "express": "^4.18.2",
     "express-session": "^1.17.3",
+    "helmet": "^7.1.0",
     "js-yaml": "^4.1.0",
     "sequelize": "^6.35.2",
     "socket.io": "^4.6.1",
