+++ views/pages/networks.ejs
@@ -48,7 +48,6 @@
                     <table class="table table-vcenter table-mobile-md card-table">
                       <tbody>
 
-                        <%- user_list %>
                       
                       </tbody>
                     </table>
