+++ views/pages/volumes.ejs
@@ -48,7 +48,7 @@
                     <table class="table table-vcenter table-mobile-md card-table">
                       <tbody>
 
-                        <%- user_list %>
+                        
                       
                       </tbody>
                     </table>
